feat: 입찰 스팸 차단 시 실제 requestCount를 로그에 저장하도록 수정 및 long으로 타입 변경 - (#314)

+++ detector/src/main/java/com/back/detector/domain/BidSpamLog.java
@@ -22,10 +22,10 @@ public class BidSpamLog extends BaseTimeEntity {
     private Long userId;
 
     @Column(nullable = false)
-    private int requestCount;
+    private long requestCount;
 
     @Column(nullable = false)
-    private int timeWindowMinutes;
+    private long timeWindowMinutes;
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
